services/order-api/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile '.\requirements.in'
#
blinker==1.9.0
    # via flask
click==8.3.1
    # via flask
colorama==0.4.6
    # via click
flask==3.1.3
    # via -r requirements.in
itsdangerous==2.2.0
    # via flask
jinja2==3.1.6
    # via flask
kafka-python==2.3.0
    # via -r requirements.in
markupsafe==3.0.3
    # via
    #   flask
    #   jinja2
    #   werkzeug
werkzeug==3.1.6
    # via flask
